Adds plus-minus receipt reader.
This was part of a hiring quiz some time ago. It should be sufficiently
anonymized to not reveal the source. Data files and instructions are
excluded, for that reason. The project was identified as a "script",
which I took to mean that it should not be production/release quality.

Feedback on the code:
- Didn't like use of money library.
- Didn't like use of transient. (I chose to use this because it was an
isolated collection constructor.)
- Submission didn't include test files (which I maintained in a scratch
file which is now lost).